[esm-integration] Fix acorn optimizer to handle ESM imports/exports
Closure compiler still can't handle these sadly.
